refactor(audit): type the read response strictly + dedupe event vocabulary
Erik review: the audit read type (ApiAuditTrailEvent) was loose (resource/
action/outcome as plain strings). Type it strictly to the current schema
version so consumers get exhaustiveness. To avoid duplicating the literals that
already live in @nangohq/audit, move the shared vocabulary (actor/target types,
outcome, resource, action, version + actor/target/context shapes) into
@nangohq/types as the single source of truth; @nangohq/audit re-exports it, so
the emit side (AuditEvent) and the read side share one definition. metadata
stays loose and version is a literal discriminant — the version-discriminated
union / transform-to-latest + runtime validation is deferred to the contract-
versioning work.
