+++ packages/types/lib/index.ts
@@ -2,6 +2,7 @@ export type * from './db.js';
 export type * from './api.js';
 export type * from './api.endpoints.js';
 export type * from './audit-trail/api.js';
+export type * from './audit-trail/event.js';
 
 export type * from './gettingStarted/db.js';
 export type * from './gettingStarted/dto.js';
